requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiosql==3.0.0
arrow==0.15.5             # via withings-api
bbcode==1.0.33
cachetools==4.1.1         # via google-auth
certifi==2019.11.28       # via requests
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via flask
cryptography==2.8         # via pyopenssl
dropbox==9.4.0
fitbit==0.3.1
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask==1.1.1
geographiclib==1.50       # via geopy
geopy==2.0.0
google-api-core==1.22.1   # via google-api-python-client
google-api-python-client==1.10.0
google-auth-httplib2==0.0.4  # via google-api-python-client
google-auth-oauthlib==0.4.1
google-auth==1.20.1
googleapis-common-protos==1.52.0  # via google-api-core
googlemaps==4.4.1
gpxpy==1.4.2
gunicorn==20.0.4
html-slacker==0.1.6
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10.3
markupsafe==1.1.1         # via jinja2
migra[pg]==1.0.1575954682
oauthlib==3.1.0           # via requests-oauthlib
pathlib==1.0.1            # via sqlbag
pendulum==2.0.5
pillow==6.2.1
polar-accesslink==0.0.5
protobuf==3.13.0          # via google-api-core, googleapis-common-protos
psycopg2-binary==2.8.4
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.19           # via cffi
pyjwt==1.7.1              # via flask-jwt-extended
pyopenssl==19.1.0
python-dateutil==2.8.1    # via arrow, fitbit, pendulum
python-dotenv==0.10.3
pytz==2019.3
pytzdata==2019.3          # via pendulum
requests-oauth==0.4.1     # via withings-api
requests-oauthlib==1.3.0  # via fitbit, google-auth-oauthlib, withings-api
requests==2.22.0
rsa==4.6                  # via google-auth
schedule==0.6.0
schemainspect==0.1.1576717677  # via migra
six==1.13.0               # via cryptography, dropbox, flask-cors, flask-jwt-extended, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, migra, protobuf, pyopenssl, python-dateutil, schemainspect, slackclient, sqlbag, websocket-client
slackclient==1.3.1
sqlalchemy==1.3.12        # via schemainspect, sqlbag
sqlbag==0.1.1579049654
staticmap==0.5.4
uritemplate==3.0.1        # via google-api-python-client
urllib3==1.25.7           # via requests
websocket-client==0.54.0
werkzeug==0.16.0
withings-api==2.1.4

# The following packages are considered to be unsafe in a requirements file:
# setuptools
